+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 <!--next-version-placeholder-->
-
+- handle  & in the print to terminal
 ## v0.1.0 (2023-01-18)
 
 ### Feature